Splitting authorization concern
